pyproject.toml
[tool.poetry]
name = "llm_difficulty_recommendations"
version = "0.1.0"
description = "A research project consisting of 1. assessing the difficulty of a French text. 2. Categorising a text according to its subject. 3. Recommending texts according to their difficulty and subject."
authors = ["Anonymous for peer review"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.9.13"
requests = "^2.31.0"
datasets = "^2.14.5"
huggingface-hub = "^0.16.4"
kaggle = "^1.5.16"
jupyter = "^1.0.0"
scikit-learn = "^1.3.0"
tiktoken = "^0.4.0"
gitpython = "^3.1.36"
numpy = "^1.25.2"
matplotlib = "^3.7.3"
rich = "^13.6.0"
wheel = "^0.41.2"
mysql-connector-python = "^8.1.0"
ray = {extras = ["data"], version = "^2.10.0"}
transformers = "^4.34.0"
accelerate = "^0.23.0"
bitsandbytes = "^0.41.1"
seaborn = "^0.13.0"
peft = "^0.5.0"
torch = "^2.1.0"
torchvision = "^0.16.0"
torchaudio = "^2.1.0"
slurmray = "^3.1.6"
modcell = "^0.1.1"
sacremoses = "^0.1.1"
black = "^23.11.0"
textstat = "^0.7.3"
tensorboard = "^2.15.1"
torch-tb-profiler = "^0.4.3"
sentencepiece = "^0.1.99"
wandb = "^0.16.0"
deepspeed = "^0.12.6"
packaging = "^23.2"
safetensors = "^0.4.2"
sentence-transformers = "^2.3.1"
nltk = "^3.8.1"
rouge-score = "^0.1.2"
pandas = "2.1.4"
colorama = "^0.4.6"
openai = "^1.54.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
